Unsigned ECS component and system indexes.

+++ src/ecs/f_entity.c
@@ -37,13 +37,13 @@ struct FEntity {
     FComponentInstance* componentsTable[F_CONFIG_ECS_COM_NUM]; // Comp, or NULL
 };
 
-static FComponentInstance* componentAdd(FEntity* Entity, int ComponentIndex, const FComponent* Component, const void* TemplateData)
+static FComponentInstance* componentAdd(FEntity* Entity, unsigned ComponentIndex, const FComponent* Component, const void* TemplateData)
 {
     FComponentInstance* c = f_component__instanceNew(
                                 Component, Entity, TemplateData);
 
     Entity->componentsTable[ComponentIndex] = c;
-    f_bitfield_set(Entity->componentBits, (unsigned)ComponentIndex);
+    f_bitfield_set(Entity->componentBits, ComponentIndex);
 
     return c;
 }
@@ -93,7 +93,7 @@ FEntity* f_entity_new(const char* Template, const void* Context)
         e->id = f_str_dup(id);
         e->template = template;
 
-        for(int c = F_CONFIG_ECS_COM_NUM; c--; ) {
+        for(unsigned c = F_CONFIG_ECS_COM_NUM; c--; ) {
             if(f_template__componentHas(template, c)) {
                 componentAdd(e,
                              c,
@@ -356,7 +356,7 @@ void f_entity_activeSetPermanent(FEntity* Entity)
     F_FLAGS_SET(Entity->flags, F_ENTITY__ACTIVE_PERMANENT);
 }
 
-void* f_entity_componentAdd(FEntity* Entity, int ComponentIndex)
+void* f_entity_componentAdd(FEntity* Entity, unsigned ComponentIndex)
 {
     const FComponent* component = f_component__get(ComponentIndex);
 
@@ -383,7 +383,7 @@ void* f_entity_componentAdd(FEntity* Entity, int ComponentIndex)
     return componentAdd(Entity, ComponentIndex, component, NULL)->buffer;
 }
 
-bool f_entity_componentHas(const FEntity* Entity, int ComponentIndex)
+bool f_entity_componentHas(const FEntity* Entity, unsigned ComponentIndex)
 {
     #if F_CONFIG_BUILD_DEBUG
         f_component__get(ComponentIndex);
@@ -392,7 +392,7 @@ bool f_entity_componentHas(const FEntity* Entity, int ComponentIndex)
     return Entity->componentsTable[ComponentIndex] != NULL;
 }
 
-void* f_entity_componentGet(const FEntity* Entity, int ComponentIndex)
+void* f_entity_componentGet(const FEntity* Entity, unsigned ComponentIndex)
 {
     #if F_CONFIG_BUILD_DEBUG
         f_component__get(ComponentIndex);
@@ -403,7 +403,7 @@ void* f_entity_componentGet(const FEntity* Entity, int ComponentIndex)
     return instance ? instance->buffer : NULL;
 }
 
-void* f_entity_componentReq(const FEntity* Entity, int ComponentIndex)
+void* f_entity_componentReq(const FEntity* Entity, unsigned ComponentIndex)
 {
     #if F_CONFIG_BUILD_DEBUG
         f_component__get(ComponentIndex);
